consensus: chain specific logic extracted
